# docs/api/openapi.yaml
openapi: 3.1.0
...
info:
  title: Uptime Cairn API
  version: 1.0.0-draft
  summary: All-in-one open source uptime monitoring and reporting platform.
  description: |
    The complete public surface of Uptime Cairn. The dashboard is the first
    client of this API, never a privileged one — there are no internal
    endpoints the UI can reach that an API key cannot.

    ## Authentication

    Two schemes, either of which authenticates any non-public operation:

    - `ApiKeyAuth` — `Authorization: Bearer cairn_<key>`. Scoped, expiring,
      revocable. The scheme for automation, CI/CD, and Terraform.
    - `SessionCookie` — the browser session the dashboard uses after
      `POST /api/v1/auth/login`. Cookie-authenticated state-changing requests
      additionally require a CSRF token in `X-Cairn-CSRF-Token`.

    Operations tagged `Public` require no authentication: the status page read
    path, subscriber confirmation, and push-monitor ingest.

    ## Pagination

    Every list endpoint paginates by opaque cursor, keyed on `(updated_at, id)`,
    per ADR-004 — uniformly, with no exception for installs small enough that
    sending everything would "work". Offsets are unstable against rows that
    change while a user is paginating, and status changes constantly.

    Cursor responses carry no total count. A count for the current filter comes
    from `GET /api/v1/monitors/membership`, which a client polling for live-view
    freshness is already calling.

...

    ## Live updates
...
servers:
...
security:
  - ApiKeyAuth: []
  - SessionCookie: []
...
paths:
...
  /api/v1/monitors:
    get:
      tags: [Monitors]
      operationId: listMonitors
      summary: List monitors
      description: |
        Server-side paginated, filtered, sorted, and searched. This endpoint must remain
        responsive with 5,000 monitors configured — it is covered by the CI load-test gate.

        Filters combine with AND. Repeated array parameters combine with OR within the
        parameter, so `tag_id=a&tag_id=b` matches monitors carrying either tag.
      x-cairn-phase: 1
      x-cairn-scopes: [monitors:read]
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/CollectionLimit"
        - $ref: "#/components/parameters/Search"
        - name: sort
          in: query
          description: |
            Result ordering. Constrained to the cursor key, because ADR-004 fixes
            cursor pagination on `(updated_at, id)` and a keyset cursor can only
            paginate the ordering it is keyed on.

            Sorting by name, status, or uptime is therefore **not offered** — see the
            note in docs/api/README.md. Sorting a page client-side would order 25 rows
            out of 5,000, which is worse than not offering it.
          schema:
            type: string
            default: -updated_at
            enum: [updated_at, -updated_at]
        - name: status
          in: query
          description: Filter by current status.
          schema:
            type: array
            items: { $ref: "#/components/schemas/MonitorStatus" }
          style: form
          explode: true
        - name: type
          in: query
          description: Filter by monitor type.
          schema:
            type: array
            items: { $ref: "#/components/schemas/MonitorType" }
          style: form
          explode: true
        - name: tag_id
          in: query
          schema:
            type: array
            items: { type: string, format: uuid }
          style: form
          explode: true
        - name: group_id
          in: query
          schema:
            type: array
            items: { type: string, format: uuid }
          style: form
          explode: true
        - name: enabled
          in: query
          schema: { type: boolean }
        - name: include
          in: query
          description: |
            Optionally embed related data. `last_heartbeat` and `uptime` cost extra
            per-row work; the dashboard list view requests them, exports do not.

            `heartbeats` embeds a bounded run of recent checks per monitor — the
            strip a list view draws under each row. It is resolved for the whole
            page in one statement rather than one request per row, which is the
            fan-out this parameter exists to prevent: a client that requests it
            per row has reintroduced exactly the cost the 5,000-monitor gate
            measures. Its length is `heartbeats_limit`, and it is bounded on
            purpose so the embed's cost tracks the viewport rather than the
            install (ADR-004).
          schema:
            type: array
            items: { type: string, enum: [last_heartbeat, heartbeats, uptime, tags, group] }
          style: form
          explode: false
        - name: heartbeats_limit
          in: query
          description: |
            How many recent checks `include=heartbeats` returns per monitor.
            Clamped to the maximum rather than refused, the same as `limit`.
          schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
      responses:
        "200":
          description: A page of monitors.
          content:
            application/json:
              schema:
                type: object
                required: [data, pagination]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Monitor" }
                  pagination: { $ref: "#/components/schemas/CursorPagination" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [Monitors]
      operationId: createMonitor
      summary: Create a monitor
      description: |
        The monitor begins checking immediately unless `enabled` is false. Its first
        heartbeat is recorded on the next scheduler tick, not synchronously with this call.
      x-cairn-phase: 1
      x-cairn-scopes: [monitors:write]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MonitorWrite" }
      responses:
        "201":
          description: Monitor created.
          headers:
            Location:
              description: URL of the created monitor.
              schema: { type: string, format: uri-reference }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Monitor" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
  /api/v1/monitors/{monitorId}:
    parameters:
      - $ref: "#/components/parameters/MonitorId"
    get:
      tags: [Monitors]
      operationId: getMonitor
      summary: Retrieve a monitor
      x-cairn-phase: 1
      x-cairn-scopes: [monitors:read]
      parameters:
        - name: include
          in: query
          schema:
            type: array
            items: { type: string, enum: [last_heartbeat, heartbeats, uptime, tags, group, certificate] }
          style: form
          explode: false
        - name: heartbeats_limit
          in: query
          description: How many recent checks `include=heartbeats` returns. Clamped, not refused.
          schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
      responses:
        "200":
          description: The monitor.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Monitor" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [Monitors]
      operationId: updateMonitor
      summary: Update a monitor
      description: |
        Partial update. Omitted fields are left unchanged; explicit `null` clears a
        nullable field. A monitor's `type` is immutable — changing check semantics
        means creating a new monitor, so that its history stays meaningful.
      x-cairn-phase: 1
      x-cairn-scopes: [monitors:write]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MonitorUpdate" }
      responses:
        "200":
          description: The updated monitor.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Monitor" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
    delete:
      tags: [Monitors]
      operationId: deleteMonitor
      summary: Delete a monitor
      description: |
        Deletes the monitor and its history. Irreversible. Monitors that are the parent
        of a dependency relationship cannot be deleted until their children are reassigned.
      x-cairn-phase: 1
      x-cairn-scopes: [monitors:write]
      responses:
        "204": { description: Monitor deleted. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
...
  /api/v1/monitors/{monitorId}/heartbeats:
    parameters:
      - $ref: "#/components/parameters/MonitorId"
    get:
      tags: [Monitor History]
      operationId: listHeartbeats
      summary: List raw heartbeats
      description: |
        Individual check results, newest first, within a time range. Raw heartbeats are
        retained per the configured retention policy; older ranges are served from
        rollups via `/history` instead. A range that falls wholly outside raw retention
        returns 200 with an empty page and a `Warning` header, not an error.
      x-cairn-phase: 1
      x-cairn-scopes: [heartbeats:read]
      parameters:
        - $ref: "#/components/parameters/From"
        - $ref: "#/components/parameters/To"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - name: important_only
          in: query
          description: Return only state-change heartbeats — the events, not every tick.
          schema: { type: boolean, default: false }
        - name: status
          in: query
          schema:
            type: array
            items: { $ref: "#/components/schemas/HeartbeatStatus" }
          style: form
          explode: true
      responses:
        "200":
          description: A page of heartbeats.
          content:
            application/json:
              schema:
                type: object
                required: [data, pagination]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Heartbeat" }
                  pagination: { $ref: "#/components/schemas/CursorPagination" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: cairn_<key>
      description: |
        A scoped API key: `Authorization: Bearer cairn_<key>`. Keys carry an explicit
        scope list, an optional expiry, and a last-used timestamp, and can be revoked
        at any time. Every operation names the scopes it requires in `x-cairn-scopes`.
    SessionCookie:
      type: apiKey
      in: cookie
      name: cairn_session
      description: |
        The browser session established by `POST /api/v1/auth/login`. The cookie is
        `HttpOnly`, `SameSite=Lax`, and `Secure` when served over HTTPS.
        State-changing requests authenticated this way must also carry a matching
        `X-Cairn-CSRF-Token` header.
  parameters:
    CollectionLimit:
      name: limit
      in: query
      description: |
        Maximum items to return. The server caps this; requesting more is clamped,
        not rejected.
      schema: { type: integer, minimum: 1, maximum: 200, default: 25 }
    Search:
      name: q
      in: query
      description: Free-text search across the resource's human-readable fields.
      schema: { type: string, maxLength: 200 }
    Cursor:
      name: cursor
      in: query
      description: |
        Opaque cursor from a previous response's `pagination.next_cursor`. Cursors are
        not decodable, not sortable, and not guaranteed stable across releases.
      schema: { type: string }
...
    MonitorId:
      name: monitorId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: The request was malformed — unparseable parameters, an invalid range, a bad cursor.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Unauthorized:
      description: No credentials were supplied, or they were not valid.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
  schemas:
...
    MonitorType:
      type: string
      description: |
        The check a monitor performs. Immutable after creation — changing it would make
        the monitor's own history incomparable with itself.
      enum: [http, tcp, icmp, dns, tls_expiry, domain_expiry, push, docker, grpc]
    MonitorStatus:
      type: string
      description: |
        `pending` covers both a monitor that has never reported and one that is retrying
        below its failure threshold — it has not yet earned a verdict either way.
      enum: [up, down, pending, paused, maintenance]
...